+++ src/app/admin/components/admin-product-form/admin-product-form.component.ts
@@ -1,49 +1,51 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
+import { Store } from '@ngrx/store';
+import { Subscription } from 'rxjs';
+import { selectSelectedProductByUrl } from 'src/app/core/@ngrx';
 
-import { ProductsPromiseService } from 'src/app/product-list';
 import { Category, ProductModel } from 'src/app/shared';
+import * as ProductsActions from '../../../core/@ngrx/products/products.actions';
+import * as RouterActions from '../../../core/@ngrx/router/router.actions';
+import { AutoUnsubscribe } from 'src/app/core/decorators';
 
 @Component({
   selector: 'app-admin-product-form',
   templateUrl: './admin-product-form.component.html',
   styleUrls: ['./admin-product-form.component.css']
 })
+@AutoUnsubscribe()
 export class AdminProductFormComponent implements OnInit {
 
   @Input() productID!: string;
   
   product!: ProductModel;
   categories = Object.values(Category);
+  private sub!: Subscription;
 
   constructor(
-    private productService: ProductsPromiseService,
-    private router: Router
+    private store: Store
   ) { }
 
   ngOnInit(): void {
     this.product = new ProductModel();
 
-    if (this.productID) {
-      this.productService.getProduct(+this.productID)
-        .then((product: ProductModel | undefined) => {
-          this.product = product ?? new ProductModel();
-        })
-        .catch(err => console.log(err));
-    }
+    this.sub = this.store.select(selectSelectedProductByUrl)
+      .subscribe((product: ProductModel) => {
+        this.product = {...product};
+      });
   }
 
   onSaveProduct(): void {
     const product = { ...this.product } as ProductModel;
 
-    const method = product.id ? 'updateProduct' : 'createProduct';
-    this.productService[method](this.product)
-      .then(() => this.onGoBack())
-      .catch(err => console.log(err));
+    if (product.id) {
+      this.store.dispatch(ProductsActions.updateProduct({ product }));
+    } else {
+      this.store.dispatch(ProductsActions.createProduct({ product }));
+    }
   }
 
   onGoBack(): void {
-    this.router.navigate(['admin/products']);
+    this.store.dispatch(RouterActions.back());
   }
-
 }