+++ src/app/core/@ngrx/app.state.ts
@@ -1,0 +1,7 @@
+import { ProductsState } from './products';
+
+export const productsFeatureKey = "products";
+
+export interface AppState {
+    [productsFeatureKey]: ProductsState
+}