+++ src/app/core/@ngrx/products/products.state.ts
@@ -1,0 +1,16 @@
+import { ProductModel } from "src/app/shared";
+
+export interface ProductsState {
+    data: ReadonlyArray<ProductModel>;
+    loading: boolean;
+    loaded: boolean;
+    error: Error | string | null;
+  }
+  
+  export const initialProductsState: ProductsState = {
+      data: [],
+      loading: false,
+      loaded: false,
+      error: null
+  };
+