+++ src/app/core/guards/can-activate-auth.guard.ts
@@ -1,14 +1,17 @@
 import { inject } from '@angular/core';
-import { CanActivateFn, Router } from '@angular/router';
+import { CanActivateFn } from '@angular/router';
+import { Store } from '@ngrx/store';
+
 import { AuthService } from '..';
+import * as RouterActions from '../@ngrx/router/router.actions';
 
 export const canActivateAuthGuard: CanActivateFn = (route, state) => {
   const authService = inject(AuthService);
-  const router = inject(Router)
+  const store = inject(Store);
   if (authService.checkLogin(state.url) && authService.isAdmin) {
     return true;
   } else {
-    router.navigate(['/login']);
+    store.dispatch(RouterActions.go({ path: ['/login'] }));
     return false;
   }
 };