+++ src/app/core/services/auth.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Observable, delay, of, tap } from 'rxjs';
+import { LocalStorageService } from './local-storage.service';
 
 @Injectable({
   providedIn: 'root'
@@ -20,7 +21,7 @@ export class AuthService {
   isLoggedIn: boolean = false;
   redirectUrl: string = '';
 
-  constructor() { }
+  constructor(private localStorageService: LocalStorageService) { }
 
   login(role: string): Observable<boolean> {
     return of(true).pipe(
@@ -28,6 +29,7 @@ export class AuthService {
       tap(val => {
         this.isLoggedIn = val;
         this.isAdmin = role === "admin";
+        this.persistLogin();
       })
     );
   }
@@ -35,10 +37,18 @@ export class AuthService {
   logout(): void {
     this.isLoggedIn = false;
     this.isAdmin = false;
+    this.persistLogin();
   }
 
   checkLogin(url: string): boolean {
     this.redirectUrl = url;
+    this.isLoggedIn = this.localStorageService.getItem('isLoggedIn') ?? false;
+    this.isAdmin = this.localStorageService.getItem('isAdmin') ?? false;
     return this.isLoggedIn;
+  }
+
+  private persistLogin(): void {
+    this.localStorageService.setItem('isLoggedIn', this.isLoggedIn);
+    this.localStorageService.setItem('isAdmin', this.isAdmin);
   }
 }